add server struct
